src/requirements.txt
langchain-aws==0.2.7
langchain-core==0.3.63
python-dotenv==1.0.1
boto3==1.35.0
botocore==1.35.0
langgraph==0.2.40
pydantic==2.10.4
